requirements.txt
# client-side requirements, pretty light-weight right?
numpy
pyzmq >= 17.1.0  # python zmq
bert-serving-client
Flask
flask_json
